pyproject.toml
[project]
name = "etchant"
version = "0.1.0"
description = "AI-powered PCB design agent generating KiCad projects from natural language specs"
requires-python = ">=3.11"
dependencies = [
    "skidl>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
]
kicad = [
    "pyspice>=1.5",
]

[project.scripts]
etchant = "etchant.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["etchant"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "requires_skidl: tests that require SKiDL and KiCad libraries installed",
]

[tool.coverage.run]
source = ["etchant"]

[tool.coverage.report]
fail_under = 80
show_missing = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
